Improved UI and added SelectableButton
Added getModulesByType to App.

Clickable now sets the cursor to HAND when hovered for a better UX.

Added afterLoadModules function to module class, this is fired after all modules are loaded.

Added mouseMoved function to module class, this is fired when the mouse is moved.

Added SelectableButton class. This can be used as toggle button when there is only one button in the group. Else it will behave like a radio button.

NOTE: Single SelectableButtons don't work yet.