Finished startscreen
Button's can now be disabled. This will prevent the callback from being called and the button will get a grayscale version of its color.

The PlayerButton now actually sets the player variables when the UI is updated. Also fixed location of the button.

Fixed stupid mistake in PlayerManager.

Fixed StartScreen start button. The screen now also validates the configuration and only enables the button when everything is filled in correctly.

TextInput now allows you to specify a callback.

Added colorGrayscale and parseValue functions to utils. Check the function comments for more info.